pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "polymarket-bot"
version = "0.1.0"
requires-python = ">=3.11"
dependencies = [
    "httpx>=0.27",
    "websockets>=12",
    "pyyaml>=6",
    "click>=8",
    "structlog>=24",
    "aiosqlite>=0.20",
    "python-dotenv>=1.0",
    "eth-account>=0.11",
    "web3>=6",
    "aiohttp>=3.9",
    "python-telegram-bot>=21",
]

[project.scripts]
polybot = "cli.main:cli"

[tool.setuptools.packages.find]
where = ["."]
include = ["src*", "cli*"]
